+++ research.py
@@ -67,10 +67,12 @@ def main(des):
 
     def show_att_data():
         show_d = Tk()
-        show_d.geometry("640x480")
-        Label(show_d, text="Student Name").grid(row=0, column=1)
-        Label(show_d, text="Att time").grid(row=0, column=2)
-        Label(show_d, text="Att Date").grid(row=0, column=3)
+        show_d.configure(bg="#f7da8f")
+        show_d.title("View The Attendance Data")
+        show_d.geometry("800x480")
+        Label(show_d, text="Name Of The Student's",borderwidth=5,relief="ridge",font=("Varela Round", 20),bg="yellow").grid(row=0, column=1,padx=10,pady=10)
+        Label(show_d, text="Attendance Time",borderwidth=5,relief="ridge",font=("Varela Round", 20),bg="yellow").grid(row=0, column=2,padx=10,pady=10)
+        Label(show_d, text="Attendance Date",borderwidth=5,relief="ridge",font=("Varela Round", 20),bg="yellow").grid(row=0, column=3,padx=10,pady=10)
         sql_query = "SELECT * FROM att"
         try:
             cursor.execute(sql_query)
@@ -79,12 +81,13 @@ def main(des):
             print(e)
         nextmv = 1
         for row in results:
-            Label(show_d, text=row[0]).grid(row=nextmv, column=1)
-            Label(show_d, text=row[1]).grid(row=nextmv, column=2)
-            Label(show_d, text=row[2]).grid(row=nextmv, column=3)
+            Label(show_d, text=row[0],borderwidth=2,relief="solid",font=("Varela Round", 20),bg="#bff78b",width=18).grid(row=nextmv, column=1,pady=2)
+            Label(show_d, text=row[1],borderwidth=2,relief="solid",font=("Varela Round", 20),bg="#bff78b",width=13).grid(row=nextmv, column=2,pady=2)
+            Label(show_d, text=row[2],borderwidth=2,relief="solid",font=("Varela Round", 20),bg="#bff78b",width=13).grid(row=nextmv, column=3,pady=2)
             nextmv += 1
 
     def save1(name, frame):
+        frame = cv2.resize(frame, (216,216))
         cv2.imwrite(f"my_image\{name}.jpg", frame)
 
     def register_page():
@@ -102,20 +105,32 @@ def main(des):
                 messagebox.showerror(
                     "Text Box Empty", "First Fill The Name Of Student "
                 )
-                register_page()
             else:
                 cap = cv2.VideoCapture(0)
+                line_inc=0
+                co=480
                 while True:
                     # cap.open(address)
                     _, frame = cap.read()
-                    cv2.imshow("Press 'Q' To Capture Image", frame)
-                    frame = cv2.resize(frame, (640, 480))
+                    
                     if cv2.waitKey(1) & 0xFF == ord("q"):
                         name = img_name.get()
                         save1(name, frame)
                         cv2.destroyAllWindows()
                         cap.release()
                         break
+                    
+                    
+                    cv2.line(frame,(0,line_inc),(640,line_inc),(0,255,0),2)
+                    if line_inc==0:
+                        co=480
+                    if co>=line_inc:
+                        line_inc+=10
+                    elif line_inc>co:
+                        line_inc-=10
+                        co=0    
+                    cv2.imshow("Press 'Q' To Capture Image", frame)
+                    
 
         Label(
             reg,
@@ -128,7 +143,7 @@ def main(des):
 
         Label(
             reg,
-            text="First Enter The Name Of Student :",
+            text="Enter Student Name :",
             font=("Varela Round", 20),
             bg="#f7da8f",
         ).place(x=60, y=150)
@@ -172,17 +187,29 @@ def main(des):
         def time_set():  
             set_t=Tk()
             set_t.geometry("640x480")
+            set_t.title("Set The Time For Attendance..")
+            set_t.configure(bg="#f7da8f")
             cu_tm=datetime.now().time()
             cu_tm = cu_tm.strftime("%I:%M:%S")
             hr_set=StringVar(set_t)
             mi_set=StringVar(set_t)
-            Label(set_t,text=f"Current Time :{cu_tm}").pack()
-            Label(set_t,text="Enter Hours").pack()
-            Entry(set_t,textvariable=hr_set).pack()
+            
+            Label(set_t,text=f"Current Time {cu_tm}",font=("Arial",30,"bold"),bg="orange",fg="gray",borderwidth=4,relief="raised").pack(pady=20)
+            
 
-            Label(set_t,text="Enter Minuts").pack()
-            Entry(set_t,textvariable=mi_set).pack()
-            Button(set_t,text="Set Timer",command=lambda:start(hr_set.get(),mi_set.get(),"on",set_t)).pack()
+            Label(set_t,text="Hours",font=("Arial",20),bg="#f7da8f").place(x=110,y=100)
+            Label(set_t,text="Minutes",font=("Arial",20),bg="#f7da8f").place(x=110,y=180)
+
+            Entry(set_t,textvariable=hr_set,font=("Arial",20),bg="#f7da8f",width=2,relief="sunken",borderwidth=3).place(x=230,y=100)
+            Entry(set_t,textvariable=mi_set,font=("Arial",20),bg="#f7da8f",width=2,relief="sunken",borderwidth=3).place(x=230,y=180)
+            
+            Button(set_t,text="Start Attendance",command=lambda:start(hr_set.get(),mi_set.get(),"on",set_t),
+                        font=("Abril Fatface", 20, "bold"),
+                        borderwidth=5,
+                        relief=RAISED,
+                        bg="gray",
+                        fg="white",
+        width=15).place(x=110,y=250)
         
         att = Tk()
         width = att.winfo_screenwidth()
@@ -201,44 +228,45 @@ def main(des):
         root.destroy()
 
         def start(hr_get,mi_get,st,des):
-            if des!="off" :
-                des.destroy()           
-            global img_len
-            inc=0   
-            wait = 10
-            che = 1                     
-            time = datetime.now()
-            cap = cv2.VideoCapture(0)
-            # add = "http://192.168.0.100:8080/video"
-            # cap.open(add)
-            res=[]
-            while True:
-                ontime = time.strftime("%I:%M:%S")
-                if st=="on":
-                    tm1=f"{hr_get}:{mi_get}:00"
-                    print(tm1,"\t",ontime)
-                    if ontime==tm1:
-                        cv2.destroyAllWindows()
-                        cap.release()
-                        break
-                _,frame = cap.read()
-                frame = cv2.resize(frame, (640, 480))
-                col = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
-                face = face_cap.detectMultiScale(
-                    col,
-                    scaleFactor=1.1,
-                    minNeighbors=5,
-                    minSize=(30, 30),
-                    flags=cv2.CASCADE_SCALE_IMAGE,
-                )
+            if hr_get=="" and mi_get=="":
+                messagebox.showerror("Timing Not Found Exception","Keep Fill The Hours and Minutes Text Box")
+            else:
+                if des!="off" :
+                    des.destroy()           
+                global img_len
+                inc=0                    
+                time = datetime.now()
+                cap = cv2.VideoCapture(0)
+                # add = "http://192.168.0.100:8080/video"
+                # cap.open(add)
+                res=[]
+                while True:
+                    ontime = time.strftime("%I:%M:%S")
+                    if st=="on":
+                        tm1=f"{hr_get}:{mi_get}:00"
+                        print(tm1,"\t",ontime)
+                        if ontime==tm1:
+                            cv2.destroyAllWindows()
+                            cap.release()
+                            break
+                    _,frame = cap.read()
+                    frame = cv2.resize(frame, (640, 480))
+                    col = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
+                    face = face_cap.detectMultiScale(
+                        col,
+                        scaleFactor=1.1,
+                        minNeighbors=5,
+                        minSize=(30, 30),
+                        flags=cv2.CASCADE_SCALE_IMAGE,
+                    )
 
-                if che >= wait:
+                    
                     if img_len == 0:
                         cv2.destroyAllWindows()
                         cap.release()
                         Label(
                             att,
-                            text="Success Fully Mark The Attandence Of All The Student",
+                            text="Success Fully Mark The Attendance Of All The Student",
                             borderwidth=5,
                             relief=RAISED,
                             bg="#8b5a94",
@@ -260,10 +288,9 @@ def main(des):
                             )
                             break
                     time = datetime.now()
-
                     if res!=[]:                                     
                         if res[0] == True:
-                            ondate = time.strftime("%d-%m-%Y")
+                            ondate = time.strftime("%d /%m /%Y")
                             ontime = time.strftime("%I:%M:%S")
                             if st=="on":
                                 tm1=f"{hr_get}:{mi_get}:00"
@@ -282,7 +309,7 @@ def main(des):
                                 connection.commit()
                                 cv2.putText(
                                     frame,
-                                    "Attandence Marked..",
+                                    "Attendance Marked..",
                                     (50, 50),
                                     cv2.FONT_HERSHEY_PLAIN,
                                     3,
@@ -308,8 +335,7 @@ def main(des):
                             face_enco.pop(inc)
                             inc=0
                             img_len -= 1
-                            cv2.imshow("Attandence Press 'E'  To Exit Frame", frame)
-
+                            
                         else:
                             for x, y, w, h in face:
                                 cv2.putText(
@@ -322,21 +348,21 @@ def main(des):
                                     3,
                                 )
                                 cv2.rectangle(frame, (x, y), (x + w, y + h), (0, 0, 255), 2)
-                                cv2.imshow("Attandence Press 'E'  To Exit Frame", frame)
+                                
                         inc += 1 
                         if inc == img_len:
                             inc = 0          
-
-                cv2.imshow("Attandence Press 'E'  To Exit Frame", frame)
-                if cv2.waitKey(1) & 0xFF == ord("e"):
-                    cv2.destroyAllWindows()
-                    cap.release()
-                    break
-                che += 1
+                    cv2.imshow("Attendance Press 'E'  To Exit Frame", frame)
+                    frame=cv2.resize(frame,(216,216))
+                    if cv2.waitKey(1) & 0xFF == ord("e"):
+                        cv2.destroyAllWindows()
+                        cap.release()
+                        break
+                    
 
         Label(
             att,
-            text="Give The Attandence Here....",
+            text="Give The Attendance Here....",
             font=("Helvetica", 50, "bold"),
             borderwidth=5,
             relief=RIDGE,
@@ -357,7 +383,7 @@ def main(des):
 
         Button(
             att,
-            text="Set Attandence Timer",
+            text="Set Attendance Timer",
             command=time_set,
             font=("Abril Fatface", 20, "bold"),
             borderwidth=5,
@@ -368,7 +394,7 @@ def main(des):
         ).pack(pady=20)
         Button(
             att,
-            text="View Attandence Data",
+            text="View Attendance Data",
             command=show_att_data,
             font=("Abril Fatface", 20, "bold"),
             borderwidth=5,
@@ -409,7 +435,7 @@ def main(des):
 
     Button(
         root,
-        text="Rigister Page",
+        text="Register Mode",
         command=register_page,
         font=("Abril  Fatface", 20, "bold"),
         borderwidth=5,
@@ -421,7 +447,7 @@ def main(des):
 
     Button(
         root,
-        text="Attandence Page",
+        text="Start Attendance",
         command=attandence_page,
         font=("Abril Fatface", 20, "bold"),
         borderwidth=5,
